Changed send_to_parser_dir as it was giving error if file already existed

+++ client.py
@@ -2,6 +2,7 @@ import os
 import sys
 import shutil
 from typing import Dict, Union
+from datetime import datetime
 
 
 FILE_PATH = os.path.abspath(__file__)
@@ -51,7 +52,13 @@ class Resume:
     def send_to_parser_dir(self) -> None:
         """Moves file from original location (cs-senior-project/input/) to resume parser directory"""
         logging.info(f"original file location: {self.file_location}")
-        shutil.move(self.file_location, RESUME_INPUT_DIR) # moving file to input dir
+        if  os.path.exists(os.path.join(RESUME_INPUT_DIR, self.file_name)):
+            timestamp = datetime.now().strftime("%Y%m%d%H%M%S")
+            new_destination = os.path.join(RESUME_INPUT_DIR, f"{self.file_name}_{timestamp}.pdf")
+            self.file_name = f"{self.file_name}_{timestamp}.pdf"
+            shutil.move(self.file_location, new_destination)
+        else:
+            shutil.move(self.file_location, RESUME_INPUT_DIR) # moving file to input dir
         self.file_location = os.path.join(RESUME_INPUT_DIR, self.file_name) # reflecting that file was moved to new location
         logging.info(f"new file location: {self.file_location}")
